feat: adjustthe firebase functions such that there are ready to be uploaded in firebase along Alexis's already existing functions
