+++ functions/index.js
@@ -3,6 +3,9 @@ const { setGlobalOptions } = require("firebase-functions/v2");
 const { initializeApp } = require("firebase-admin/app");
 const { getFirestore } = require("firebase-admin/firestore");
 const logger = require("firebase-functions/logger");
+const { onDocumentUpdated } = require("firebase-functions/v2/firestore");
+const { notifyStudentLessonConfirmedByTutor, notifyTutorForConfirmation, notifyStudentLessonPendingReview, notifyStudentAsTutorOfferedToTeach} = require("./lessonNotifications");
+
 
 initializeApp();
 
@@ -12,6 +15,36 @@ setGlobalOptions({
     memory: "256MiB"
 });
 
+//Send notification when a lesson is updated
+exports.lessonStatusChanged = onDocumentUpdated("lessons/{lessonId}", async (event) => {
+  const before = event.data.before.data();
+  const after = event.data.after.data();
+
+  if (before.status !== after.status) {
+    console.log(`Lesson status changed from ${before.status} to ${after.status}`);
+
+    if (before.status === "PENDING_TUTOR_CONFIRMATION" && after.status === "CONFIRMED") {
+      await notifyStudentLessonConfirmedByTutor(after);
+    }
+
+    if (before.status === "MATCHING" && after.status === "PENDING_TUTOR_CONFIRMATION") {
+      await notifyTutorForConfirmation(after);
+    }
+    // Notify student when lesson is pending review
+    if (before.status === "CONFIRMED" && after.status === "PENDING_REVIEW") {
+      await notifyStudentLessonPendingReview(after);
+    }
+  }
+
+  //Detect when an additional tutor offer to teach a requested lesson:
+  //more specifically when a tutor uid is added to the tutorUid list of a STUDENT_REQUESTED lesson
+  if (after.status === "STUDENT_REQUESTED" && before.tutorUid.length < after.tutorUid.length) {
+    console.log("Tutor added to the lesson's tutorUid.");
+    await notifyStudentAsTutorOfferedToTeach(after);
+  }
+});
+
+//Check if lessons are completed and update their status accordingly
 exports.checkCompletedLessons = onSchedule("* * * * *", async (event) => {
     try {
         const now = new Date();
@@ -98,6 +131,7 @@ exports.checkCompletedLessons = onSchedule("* * * * *", async (event) => {
     }
 });
 
+//Check if lessons are reviewed and update their status accordingly
 exports.checkReviewedLessons = onSchedule("*/30 * * * *", async (event) => {
     try {
         const now = new Date();
